Revert "Merge pull request #117 from Exygy/160970715-fix-environmental-images"
Reverting because we are doing a release of the work on the staging branch, but the environment modal/carousel work is not part of that release. This reverts commit abc5f6f8f540b5db9c46beafceba7bf5bbe6f7f8, reversing
changes made to 10272b7030214a354ef73c14d33e393272fad391.